fix(types): resolve all mypy type checking errors
Fixed 50 mypy errors across 11 files to enable pre-commit hook compliance:

Type annotation improvements:
- Added Optional types for conditional assignments (GradScaler, Stream)
- Fixed Polars type instantiation (pl.Binary() instead of pl.Binary)
- Added cast() for pl.from_arrow() return type narrowing
- Added missing return type annotations across all modules

Files fixed:
- polars_datasource.py: 14 errors → 0 (added full type annotations)
- schema.py: 13 errors → 0 (fixed Polars DataType instances)
- multi_stage_training.py: 1 error → 0 (Optional[GradScaler])
- gpu_prefetcher.py: 1 error → 0 (Optional[Stream])
- rust_io.py: 2 errors → 0 (cast to DataFrame)
- benchmark_polars_io.py: 2 errors → 0 (return type annotations)
- hcpe_converter.py: 1 error → 0 (dict type annotation)
- feature_store.py: 1 error → 0 (int cast for estimated_size)
- file_data_source.py: 1 error → 0 (np.dtype instance)

All 323 tests passing，ready for git push．

+++ src/maou/app/pre_process/feature.py
@@ -65,9 +65,7 @@ def make_feature_from_board_state(
             "board_id_positions must have shape (9, 9)"
         )
     if pieces_in_hand.shape != (14,):
-        raise ValueError(
-            "pieces_in_hand must have shape (14,)"
-        )
+        raise ValueError("pieces_in_hand must have shape (14,)")
 
     offset = len(PieceId) - 1
 
@@ -110,7 +108,9 @@ def make_feature_from_board_state(
                 (row > PieceId.EMPTY.value) & (row <= offset)
             )
             white_count = np.count_nonzero(row > offset)
-            weighted_score += rank_weight * (black_count - white_count)
+            weighted_score += int(
+                rank_weight * (black_count - white_count)
+            )
         current_is_black = weighted_score >= 0
 
     for rank in range(9):
